requirements.txt
google-api-python-client>=2.0
google-auth>=2.0
google-auth-httplib2
python-dotenv
requests
